+++ Sources/Sheeeeeeeeet/ActionSheet/Menu/Items/LinkItem+ActionSheet.swift
@@ -11,7 +11,7 @@ import UIKit
 extension LinkItem {
 
     open override func actionSheetCell(for tableView: UITableView) -> ActionSheetItemCell {
-        ActionSheetLinkItemCell(style: .value1)
+        ActionSheetLinkItemCell(style: preferredActionSheetCellStyle)
     }
     
     open override var actionSheetCellType: ActionSheetItemCell.Type {