+++ Sources/Sheeeeeeeeet/ActionSheet/Menu/Items/MultiSelectItem+ActionSheet.swift
@@ -11,7 +11,7 @@ import UIKit
 extension MultiSelectItem {
 
     open override func actionSheetCell(for tableView: UITableView) -> ActionSheetItemCell {
-        ActionSheetMultiSelectItemCell(style: .value1)
+        ActionSheetMultiSelectItemCell(style: preferredActionSheetCellStyle)
     }
     
     open override var actionSheetCellType: ActionSheetItemCell.Type {